# perf(sourcemap): improve perf of `search_original_line_and_column`