--- a/crates/oxc_codegen/src/sourcemap_builder.rs
+++ b/crates/oxc_codegen/src/sourcemap_builder.rs
@@ -142,7 +142,7 @@ impl SourcemapBuilder {
         // This is a hot path. When building sourcemaps, typically, positions are accessed in increasing order.
         // e.g. if the last call to this function looked at position `n`, the next call will likely be for `n+1`.
         if position >= lines[idx].byte_offset_to_start_of_line {
-            let cap = (idx + 20).min(lines.len() - 1);
+            let cap = (idx + 16).min(lines.len() - 1);
             while idx + 1 < cap && lines[idx + 1].byte_offset_to_start_of_line <= position {
                 idx += 1;
             }
@@ -152,7 +152,7 @@ impl SourcemapBuilder {
                     .saturating_sub(1);
             }
         } else {
-            let cap = idx.saturating_sub(20);
+            let cap = idx.saturating_sub(16);
             while idx > cap && lines[idx].byte_offset_to_start_of_line > position {
                 idx -= 1;
             }