@camc314 2024-12-16
there's the op for more perf gains here.

`SourceMapBuilder::tokens` is just a `Vec` with no initial size. we can probably guess the size based on the length of the text to avoid some allocations - even a rough guess should be ok.

poc:
```

Summary
  ./target/release-with-debug/examples/sourcemap ran # reserving vec capacity
    1.03 ± 0.04 times faster than ./target/release-with-debug/examples/sourcemap-with-opt1 # this PR 
    1.11 ± 0.05 times faster than ./target/release-with-debug/examples/sourcemap-main  # buiilt against main
```

https://github.com/oxc-project/oxc-sourcemap/blob/0e5ec37a00d6c563727814b9ec41232b226eff90/src/sourcemap_builder.rs#L19