@camc314 2024-12-16
> Very nice!
> 
> Did I write this code?? It looks familiar. It also looks embarrassingly messy! I'd probably write it differently if I wrote it today. I think I've got a bit better at coding over past 6 months...

yes and no i think? i think @ underfin wrote this originally, then you made some perf improvements in https://github.com/oxc-project/oxc/pull/4641

> If you've enjoyed this and you'd be interested in pushing for further gains, I have a few thoughts of other optimizations we might be able to make, which I guess I didn't see first time around.

👀 let me know what you can see