@camc314 2024-12-16
> Unrelated to this PR, but I just want to say that the conversations you two have in your PRs/issues are super helpful to somebody unfamiliar with Rust (or any low level languages).
> 
> Thanks for the explanations @overlookmotel and @camc314 for always asking questions.

❤️ this is amazing to hear. overlookmotel always has the answers

now just to make rust a part of my day job 🙂