@overlookmotel 2024-12-17
@camc314 I hope you don't mind, but I thought I'd give it a go. I've pushed another commit with an optimized implementation of `search_original_line_forwards`.

I didn't want to use `#[cold]`, but I couldn't find another way to get the assembly in the shape I wanted - fastest path for "on the same line as last time", followed by the unrolled loop covering next 16 lines.

https://godbolt.org/z/heh1cnYa4 (the unrolled loop is the part where same 3 lines repeat over and over).

Let's see what CodSpeed says!