@overlookmotel 2024-12-17
OK, the perf gain is back to where it was originally, and we don't have the problem of pathologically bad perf if code has moved a long distance any more.

If you think my code looks OK, would you like to apply the same scheme to the backwards search?